refactor: drizzleスキーマをドメイン毎にファイル分割
単一の schema.ts を schema/ ディレクトリ配下にドメイン単位で分割する。

- schema/user.ts: users, activeUsers, bannedUsers, leavedUsers とリレーション
- schema/article.ts: articles, readHistories, skippedArticles とリレーション
- schema/datetime.ts: 共通の dateTime カスタム型と normalizeDateTime
- schema/index.ts: 既存の import パスを維持するバレル

import パス `@/infrastructure/drizzle-orm/schema` は変更不要。
drizzle-kit generate でスキーマ差分なしを確認済み。
