+++ application/drizzle.config.ts
@@ -2,6 +2,6 @@ import { defineConfig } from 'drizzle-kit'
 
 export default defineConfig({
   dialect: 'sqlite',
-  schema: './src/infrastructure/drizzle-orm/schema.ts',
+  schema: './src/infrastructure/drizzle-orm/schema/index.ts',
   out: './migrations',
 })
